Move all test extensions to the same module.
Some were in test_base.py, others in utils.py.
